requirements.txt
streamlit>=1.31
google-generativeai>=0.6
python-dotenv>=1.0
pydantic>=2.0
requests>=2.31
